requirements.txt
discord.py==2.4.0
Flask==3.0.3
itsdangerous==2.2.0
Werkzeug==3.0.4
Jinja2==3.1.4
TikTokLive==6.1.12
aiohttp==3.10.10
websockets==13.0
python-dotenv==1.0.1